add upcase func
